+++ .gitignore
@@ -35,6 +35,3 @@ out/
 
 ### VS Code ###
 .vscode/
-
-# test 폴더를 전체 제외
-/test/
