+++ src/main/java/edu/du/academic_management_system/AcademicManagementSystemApplication.java
@@ -1,6 +1,8 @@
 package edu.du.academic_management_system;
 
+import edu.du.academic_management_system.config.SecurityConfig;
 import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
