+++ src/main/java/edu/du/academic_management_system/config/SecurityConfig.java
@@ -13,6 +13,8 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 
 import java.util.ArrayList;
@@ -26,6 +28,12 @@ import java.util.List;
 public class SecurityConfig{
 
     private final MemberService memberService;
+
+    @Bean
+    public PasswordEncoder passwordEncoder(){
+        return new BCryptPasswordEncoder();
+    }
+
 
     @Bean
     public UserDetailsService userDetailsService() {
@@ -48,13 +56,28 @@ public class SecurityConfig{
 
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+//        http
+//                .authorizeRequests()
+//                .antMatchers("/js/**", "/css/**").permitAll() // 정적 리소스 허용
+//                .antMatchers("/login").permitAll() // 로그인 페이지 허용
+//                .anyRequest().authenticated() // 그 외 요청은 인증 필요
+//                .and()
+//                .formLogin() // 폼 로그인 활성화
+//                .loginPage("/login") // 커스텀 로그인 페이지 설정 (필요 시)
+//                .defaultSuccessUrl("/") // 로그인 성공 후 이동할 페이지
+//                .permitAll()
+//                .and()
+//                .logout()
+//                .logoutUrl("/logout")
+//                .logoutSuccessUrl("/login?logout")
+//                .permitAll();
+
         http
                 .authorizeRequests()
-                .antMatchers("/js/**", "/css/**").permitAll()
-                .antMatchers("/login").permitAll()
-//                .antMatchers("/**").permitAll()
-                .anyRequest().authenticated();
-        http.csrf().disable();
+                .anyRequest().permitAll() // 모든 요청을 허용
+                .and()
+                .csrf().disable(); // CSRF 보호를 비활성화
+
         return http.build();
     }
 }
